refactor: split indicator registry into data + logic files (#405)
Move the INDICATOR_REGISTRY array to indicator-descriptors.ts (pure
config data). indicator-registry.ts keeps types + helpers and
re-exports the array — zero import changes across the codebase.
